VisionPath/enet/train&result/v1/requirement.txt
torch 
torchvision 
opencv-python 
numpy 
pillow
gtts
pygame